Moved excel parsing functions to mock-data. Added script for testing
