+++ script/test
@@ -1,0 +1,18 @@
+#!/bin/sh
+
+# script/test: Do everything related to testing
+
+set -e
+
+cd "$(dirname "$0")/.."
+
+echo "==> Linting..."
+eslint .
+echo "==> Done!"
+
+echo "==> Setting Up Test Database..."
+node models/setup.js
+echo "==> Done!"
+
+echo "==> Running tests..."
+jasmine
